ft-#81: Add banks endpoints
